+++ src/main/java/team7/inplace/video/application/VideoService.java
@@ -6,8 +6,8 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import team7.inplace.influencer.entity.Influencer;
 import team7.inplace.influencer.repository.InfluencerRepository;
+import team7.inplace.place.application.dto.PlaceForVideo;
 import team7.inplace.place.domain.Place;
-import team7.inplace.place.dto.PlaceForVideo;
 import team7.inplace.video.application.dto.VideoInfo;
 import team7.inplace.video.domain.Video;
 import team7.inplace.video.persistence.VideoRepository;
